Small files location refactoring
