+++ code/orders-infrastructure/src/main/java/com/academy/orders/infrastructure/accountV2/entity/AccountV2Entity.java
@@ -1,0 +1,5 @@
+package com.academy.orders.infrastructure.accountV2.entity;
+
+public class AccountV2Entity {
+}
+
