+++ code/orders-infrastructure/src/main/java/com/academy/orders/infrastructure/accountV2/repository/AccountV2JpaAdapter.java
@@ -1,0 +1,4 @@
+package com.academy.orders.infrastructure.accountV2.repository;
+
+public interface AccountV2JpaAdapter {
+}
