+++ code/orders-infrastructure/src/main/java/com/academy/orders/infrastructure/orderV2/entity/OrderV2Entity.java
@@ -1,0 +1,4 @@
+package com.academy.orders.infrastructure.orderV2.entity;
+
+public class OrderV2Entity {
+}
