+++ code/orders-infrastructure/src/main/java/com/academy/orders/infrastructure/postaddress/PostAddressV2Mapper.java
@@ -1,4 +1,4 @@
-package com.academy.orders.infrastructure.orderV2;
+package com.academy.orders.infrastructure.postaddress;
 
 import com.academy.orders.domain.postaddress.entity.PostAddressV2;
 import com.academy.orders.infrastructure.order.entity.PostAddressEntity;
