+++ code/orders-infrastructure/src/main/java/com/academy/orders/infrastructure/postaddress/entity/PostAddressV2Entity.java
@@ -1,0 +1,4 @@
+package com.academy.orders.infrastructure.postaddress.entity;
+
+public class PostAddressV2Entity {
+}
